move errors to constants
