--- token/ts/src/errors.ts
@@ -1,7 +1,0 @@
-/** @TODO: docs */
-export enum TokenError {
-    ACCOUNT_NOT_FOUND = 'ACCOUNT_NOT_FOUND',
-    INVALID_ACCOUNT_OWNER = 'INVALID_ACCOUNT_OWNER',
-    INVALID_ACCOUNT_SIZE = 'INVALID_ACCOUNT_SIZE',
-    INVALID_MINT = 'INVALID_MINT',
-}
