+++ token/ts/src/index.ts
@@ -2,4 +2,3 @@ export * from './instructions';
 export * from './state';
 export * from './transactions';
 export * from './constants';
-export * from './errors';
